Accounting for single width/height cells
Changes have been made so that the parser does not count an across or down cell
which has a width/height of 1.  If a cell has a non character value to either
side or on both top and bottom, it will not be counted as an across cell or a
down cell.